A build fix
There was an error occurring during the MinGW installation step in GitHub Actions workflow. Chocolatey successfully installed MinGW v13.2.0 (the 'upgrade of mingw was successful'). The PATH was updated to include the GCC binaries. However, the installation script tried to remove a file (libpthread.dll.a) that doesn't exist in this MinGW version/build.